move dockerignore test
