+++ provider/cmd/pulumi-resource-docker/schema.json
@@ -2161,6 +2161,9 @@
                     "type": "string",
                     "description": "\nPath to use for build context. If omitted, an empty context is used."
                 },
+                "contextHash": {
+                    "type": "string"
+                },
                 "exports": {
                     "type": "array",
                     "items": {
